+++ lua/snippets/luasnip/init.lua
@@ -1,6 +1,2 @@
-local ls = require('luasnip')
-local go = require('snippets.luasnip.go')
-
-ls.add_snippets('go', go)
-
-require('luasnip/loaders/from_vscode').lazy_load()
+require('snippets.luasnip.go')
+require('luasnip.loaders.from_vscode').lazy_load()